requirements.txt
openai
python-dotenv
huggingface-hub
numpy
pandas
datasets
chromadb
langchain
tiktoken
python-dotenv
unstructured 

